+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Here is the list of the changes and features provided by the Rainbow-CLI
 
 All notable changes to Rainbow-CLI will be documented in this file.
 
+### [1.39.0] - 2018-04-08
+- RQRAINB-830: Inject official appID/appSecret in Rainbow CLI
+- RQRAINB-829: Add command configure
+
 ### [1.38.0] - 2018-19-03
 - CRRAINB-1986: Add command for adding developer role
 - CRRAINB-2095: Add options for retrieving application metrics by month and by year